Added update ability and some bugfixes
Databases can now be updated if appropriate information is sent to the
update_anime function. Although it's currently impossible to update
anime via the interactive interface, there is currently test code to
update anime via the --update flag.
Another entry was added in the creation of the "anime" table of the
database to ensure that anime may be added after the Wikipedia url field
was added. Databases made before this fix will need to be re-created or
will need the new column added.